Refactor table Column code into a separate file
