+++ astropy/table/__init__.py
@@ -1,5 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-from .table import Column, Table, TableColumns, Row, MaskedColumn
+from .column import Column, MaskedColumn
+from .table import Table, TableColumns, Row
 from .np_utils import TableMergeError
 from .operations import join, hstack, vstack
